# Optimise TextEdit theme loading