During editor load and when switching between scripts tabs, `TextEdit` was updating its cache for no reason, this was quite slow.

This PR adds two new methods to `Control`. `begin_bulk_theme_override`, which will prevent `*_theme_*_override` methods from causing a theme update, until `end_bulk_theme_override` is called.

Secondly, since we now have a dedicated code editing node `CodeEdit`, unlike before with `TextEdit`, we had to override most items to turn it into the code editor. Now there is no longer a need to do so, the theme overrides have been moved into the default editor theme.

Finally, `TextEdits` internal `text` class, now has a `is_dirty` flag to prevent invalidating the cache when not needed.

Overall, this has increased performance for a large script on load from `450ms` to `30ms`, and switching script tabs, from `1.3 seconds`  to `80ms`.

Seems to alleviate some of the issues seen in #51349