@AnilBK 2021-08-12
Your second commit "Only update TextEdit text cache when dirty." seems interesting.
Some Setters of nodes in scene/ folder, just set the property and update/regenerate the internal values again, without checking if the newly set values is same as old values.

I wonder if  we apply  condition in setters to check if new value is same as old values wherever possible, how much performance gains would be available.