@akien-mga 2021-08-12
See #49227, it seems you're reimplementing something which already exists for preventing theme updates (CC @pycbouh).