@pycbouh 2021-08-13
> See #49227, it seems you're reimplementing something which already exists for preventing theme updates (CC @pycbouh).

Similar, but not quite the same. What I've added in the linked PR prevents changes to the `Theme` resource from propagating during bulk operations. This one solves the situation where local overrides on a control are updated and cause the theme changed signal to fire, unrelated to the `Theme` resource.

Now, the only reason why the control fires that notification at all is so it can update itself (the signal is not propagated to the children, because they don't really care for the local overrides on the parent control). So this solution seems fine and not stepping on the same ground as the linked PR. Though I'm not sure if we should expose the new methods to the scripting API, as they need to be handled with care. Or at the very least a big warning is in order in the documentation.