@Chaosus 2021-08-13
Oh, seems like this fixes the https://github.com/godotengine/godot/issues/51349 - it speeds up it very much at least!